Comet Score BETA
Comet Score and Comet Score UI, made in RShiny

Note that the actual score's calculation has been redacted for security purposes. The AI model used to create the initial dataset has also not been included for the same security purposes. I DO NOT give permission to use this code or replicate it in any way.